+++ units/src/main/scala/arrays/DoubleUArray.scala
@@ -25,11 +25,14 @@ import stasiak.karol.units._
 import scala.collection.mutable._
 
 object DoubleUArray {
+	/** Creates an array of given elements */
 	def apply[U<:MUnit](elems: DoubleU[U]*) = new DoubleUArray[U](elems.map{_.value}.toArray)
 
+	/** Concatenates all arrays into a single array. */
 	def concat[U<:MUnit](arrays: DoubleUArray[U]*) = 
 		new DoubleUArray(Array.concat(arrays.map{_.underlying}: _*))
 
+	/** Copy one array to another. */
 	def copy[U<:MUnit](
 		src: DoubleUArray[U], srcPos: Int, 
 		dest: DoubleUArray[U], destPos: Int, 
@@ -37,10 +40,18 @@ object DoubleUArray {
 		Array.copy(src.underlying, srcPos, dest.underlying, destPos, length)
 	}
 
+	/** Returns an array of length 0. */
 	def empty[U<:MUnit] = new DoubleUArray[U](Array.empty[Double])
 
+	/** Returns an array that contains the results of some element computation a number of times. */
 	def fill[U<:MUnit](n: Int)(elem: =>DoubleU[U]) = {
 		new DoubleUArray(Array.fill(n)(elem.value))
+	}
+	
+	/** Returns an array that contains a constant element a number of times. */
+	def fillUniform[U<:MUnit](n: Int)(elem: DoubleU[U]) = {
+		val elemValue = elem.value
+		new DoubleUArray(Array.fill(n)(elemValue))
 	}
 	
 	def unapplySeq[U<:MUnit](arr: DoubleUArray[U]) = Some(arr)
@@ -76,4 +87,13 @@ final class DoubleUArray[U<:MUnit] private[arrays] (private[arrays] val underlyi
 		underlying(index) = elem.value
 	}
 	
+	def zip(that: DoubleUArray[U]) = {
+		val newLength = length min that.length
+		val array = new Array[Double](newLength*2)
+		(0 until newLength) foreach { i =>
+			array(i*2)     = underlying(i)
+			array(i*2 + 1) = that.underlying(i)
+		}
+		new DoubleUArray[U](array)
+	}
 }