+++ units/src/main/scala/package.scala
@@ -24,6 +24,8 @@ package stasiak.karol
 package object units {
 	import language.higherKinds
 	import language.implicitConversions
+	import stasiak.karol.units.internal.ratios._
+	import stasiak.karol.units.internal.UnitName
 	import stasiak.karol.units.internal.UnitImpl
 	import stasiak.karol.units.internal.UnitImpl._
 	import stasiak.karol.units.internal.Strings._
@@ -49,37 +51,15 @@ package object units {
 
 	/** Unit dimensionless values have, which is 1.*/
 	type _1 = TDimensionless
-	type DoubleRatio[U<:MUnit, V<:MUnit] = Conversions.DoubleRatio[U,V]
-	type IntRatio[U<:MUnit, V<:MUnit] = Conversions.IntRatio[U,V]
-	type DoubleAffineSpaceConverter[T1<:AffineSpace, T2<:AffineSpace] = AffineSpaces.DoubleAffineSpaceConverter[T1, T2]
-	type IntAffineSpaceConverter[T1<:AffineSpace, T2<:AffineSpace] = AffineSpaces.IntAffineSpaceConverter[T1, T2]
-
-	class TUnitPowerPairName[U<:TUnitPowerPair](w:String){
-		override def toString = w
-	}
-	@inline
-	implicit def implicit_unitPowerPairName[U<:TUnitPowerPair](
-		implicit stringName: TStringName[U#UnitName#Name], power:ToInt[U#Power]
-		) = new TUnitPowerPairName[U](
-		" "+stringName.toString +(
-			if(power.toInt<0)  "^("+power.toInt+")" else (if (power.toInt>1) "^"+power.toInt else "")
-		)
-	)
-	@inline
-	implicit val implicit_dimensionlessName = new UnitName[_1]("")
-	@inline
-	implicit def implicit_consUnitName[H<:TUnitPowerPair, T<:MUnit](implicit h:TUnitPowerPairName[H], t:UnitName[T]) = 
-		new UnitName[T**H](h.toString+t.toString)
-	@inline
-	implicit def implicit_singleUnitName[H<:TUnitPowerPair](implicit h:TUnitPowerPairName[H]) = 
-		new UnitName[H](h.toString)
-
+	type DoubleRatio[U<:MUnit, V<:MUnit] = stasiak.karol.units.internal.ratios.DoubleRatio[U,V]
+	type IntRatio[U<:MUnit, V<:MUnit] = stasiak.karol.units.internal.ratios.IntRatio[U,V]
+	type DoubleAffineSpaceConverter[T1<:AffineSpace, T2<:AffineSpace] = stasiak.karol.units.internal.ratios.DoubleAffineSpaceConverter[T1, T2]
+	type IntAffineSpaceConverter[T1<:AffineSpace, T2<:AffineSpace] = stasiak.karol.units.internal.ratios.IntAffineSpaceConverter[T1, T2]
 
 	@inline
 	implicit class UnitNameBuilder(override val toString:String) extends AnyVal{
 		def unitName[U<:MUnit] = new UnitName(toString)
 	}
-
 
 	@inline
 	implicit def implicit_widening[U<:MUnit](i: IntU[U]) = i.toDouble
@@ -239,162 +219,6 @@ package object units {
 		@inline
 		def atto [U<:MUnit] = new DoubleU[U](value/1000000000000000000.0)
 	}
-
-	@inline
-	implicit def implicit_oneDoubleRatio[U<:MUnit] = new DoubleRatio[U,U](1.0)
-	@inline
-	implicit def implicit_oneIntRatio[U<:MUnit] = new IntRatio[U,U](1)
-	@inline
-	implicit def implicit_intRatioEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U,V]
-		) = new IntRatio[U,V](declaredRatio.ratio)
-	@inline
-	implicit def implicit_doubleRatioEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseDoubleRatio[U,V]
-		) = new DoubleRatio[U,V](declaredRatio.ratio)
-	@inline
-	implicit def implicit_doubleRatioIntEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U,V]
-		) = new DoubleRatio[U,V](declaredRatio.ratio)
-	@inline
-	implicit def implicit_revDoubleRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseDoubleRatio[U,V]
-		) = new DoubleRatio[V,U](1.0/declaredRatio.ratio)
-	@inline
-	implicit def implicit_revIntRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U,V]
-		) = new DoubleRatio[V,U](1.0/declaredRatio.ratio)
-
-	@inline
-	implicit def implicit_squareIntRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U#Sqrt,V#Sqrt]
-		) = new IntRatio[U,V](declaredRatio.ratio * declaredRatio.ratio)
-	@inline
-	implicit def implicit_cubeIntRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U#Cbrt,V#Cbrt]
-		) = new IntRatio[U,V](declaredRatio.ratio * declaredRatio.ratio * declaredRatio.ratio)
-	@inline
-	implicit def implicit_squareDoubleRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseDoubleRatio[U#Sqrt,V#Sqrt]
-		) = new DoubleRatio[U,V](declaredRatio.ratio * declaredRatio.ratio)
-	@inline
-	implicit def implicit_cubeDoubleRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseDoubleRatio[U#Cbrt,V#Cbrt]
-		) = new DoubleRatio[U,V](declaredRatio.ratio * declaredRatio.ratio * declaredRatio.ratio)
-	@inline
-	implicit def implicit_squareDoubleIntRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U#Sqrt,V#Sqrt]
-		) = new DoubleRatio[U,V](declaredRatio.ratio * declaredRatio.ratio)
-	@inline
-	implicit def implicit_cubeDoubleIntRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U#Cbrt,V#Cbrt]
-		) = new DoubleRatio[U,V](declaredRatio.ratio * declaredRatio.ratio * declaredRatio.ratio)
-
-	@inline
-	implicit def implicit_squareRevIntRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U#Sqrt,V#Sqrt]
-		) = new DoubleRatio[V,U](1.0/(declaredRatio.ratio * declaredRatio.ratio))
-	@inline
-	implicit def implicit_cubeRevIntRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseIntRatio[U#Cbrt,V#Cbrt]
-		) = new DoubleRatio[V,U](1.0/(declaredRatio.ratio * declaredRatio.ratio * declaredRatio.ratio))
-	@inline
-	implicit def implicit_squareRevDoubleRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseDoubleRatio[U#Sqrt,V#Sqrt]
-		) = new DoubleRatio[V,U](1.0/(declaredRatio.ratio * declaredRatio.ratio))
-	@inline
-	implicit def implicit_cubeRevDoubleRatio[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:BaseDoubleRatio[U#Cbrt,V#Cbrt]
-		) = new DoubleRatio[V,U](1.0/(declaredRatio.ratio * declaredRatio.ratio * declaredRatio.ratio))
-
-	@inline
-	implicit def implicit_intRatioAliasEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U,V]
-		) = new IntRatio[U,V](1)
-	@inline
-	implicit def implicit_doubleRatioAliasEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U,V]
-		) = new DoubleRatio[U,V](1)
-	@inline
-	implicit def implicit_intRatioAliasSqEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U#Sqrt,V#Sqrt]
-		) = new IntRatio[U,V](1)
-	@inline
-	implicit def implicit_doubleRatioAliasSqEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U#Sqrt,V#Sqrt]
-		) = new DoubleRatio[U,V](1)
-	@inline
-	implicit def implicit_intRatioAliasCbEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U#Cbrt,V#Cbrt]
-		) = new IntRatio[U,V](1)
-	@inline
-	implicit def implicit_doubleRatioAliasCbEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U#Cbrt,V#Cbrt]
-		) = new DoubleRatio[U,V](1)
-
-	@inline
-	implicit def implicit_intRatioAliasRevEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U,V]
-		) = new IntRatio[V,U](1)
-	@inline
-	implicit def implicit_doubleRatioAliasRevEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U,V]
-		) = new DoubleRatio[V,U](1)
-	@inline
-	implicit def implicit_intRatioAliasSqRevEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U#Sqrt,V#Sqrt]
-		) = new IntRatio[V,U](1)
-	@inline
-	implicit def implicit_doubleRatioAliasSqRevEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U#Sqrt,V#Sqrt]
-		) = new DoubleRatio[V,U](1)
-	@inline
-	implicit def implicit_intRatioAliasCbRevEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U#Cbrt,V#Cbrt]
-		) = new IntRatio[V,U](1)
-	@inline
-	implicit def implicit_doubleRatioAliasCbRevEv[U<:MUnit, V<:MUnit](
-		implicit declaredRatio:UnitAlias[U#Cbrt,V#Cbrt]
-		) = new DoubleRatio[V,U](1)
-
-
-
-	@inline
-	implicit def implicit_zeroPowerDoubleRatio[U<:TSingleUnit, V<:MUnit] = 
-		new PowerDoubleRatio[U,P1,V,_0](1.0)
-	@inline
-	implicit def implicit_positivePowerDoubleRatio[U<:TSingleUnit, V<:MUnit, N<:TInteger](
-		implicit ev:DoubleRatio[U^P1,V], prev:PowerDoubleRatio[U,P1,V,N]
-		) = new PowerDoubleRatio[U,P1,V,Inc[N]](prev.ratio*ev.ratio)
-	@inline
-	implicit def implicit_negativePowerDoubleRatio[U<:TSingleUnit, V<:MUnit, N<:TInteger](
-		implicit ev:DoubleRatio[U^P1,V], prev:PowerDoubleRatio[U,P1,V,N]
-		) = new PowerDoubleRatio[U,P1,V,Dec[N]](prev.ratio/ev.ratio)
-
-	@inline
-	implicit def implicit_zeroPowerIntRatio[U<:TSingleUnit, V<:MUnit] = 
-		new PowerIntRatio[U,P1,V,_0](1)
-	@inline
-	implicit def implicit_positivePowerIntRatio[U<:TSingleUnit, V<:MUnit, N<:TInteger](
-		implicit ev:IntRatio[U^P1,V], prev:PowerIntRatio[U,P1,V,N]
-		) = new PowerIntRatio[U,P1,V,Inc[N]](prev.ratio*ev.ratio)
-
-	@inline
-	implicit def implicit_oneLeftIntRatio[U<:MUnit] = new LeftIntRatio[U,U,U](1)
-	@inline
-	implicit def implicit_oneRightIntRatio[U<:MUnit] = new RightIntRatio[U,U,U](1)
-	@inline
-	implicit def implicit_leftIntRatio[U<:MUnit, V<:MUnit](implicit ev: IntRatio[U,V]) = 
-		new LeftIntRatio[U,V,V](ev.ratio)
-	@inline
-	implicit def implicit_rightIntRatio[U<:MUnit, V<:MUnit](implicit ev: IntRatio[U,V]) = 
-		new RightIntRatio[U,V,V](1)
-	@inline
-	implicit def implicit_leftIntRatioRev[U<:MUnit, V<:MUnit](implicit ev: IntRatio[V,U]) = 
-		new LeftIntRatio[U,V,U](1)
-	@inline
-	implicit def implicit_rightIntRatioRev[U<:MUnit, V<:MUnit](implicit ev: IntRatio[V,U]) = 
-		new RightIntRatio[U,V,U](ev.ratio)
 		
 	@inline
 	implicit def implicit_oneDoubleAffineSpaceConverter[U<:AffineSpace] = new DoubleAffineSpaceConverter[U,U](_)